MachineLearning/linebot/requirements.txt
Flask
gunicorn
line-bot-sdk
numpy
request
Image
keras
tensorflow
